requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
djangorestframework==3.12.4
django-phonenumber-field==5.2.0
django-import-export==2.0.2
sqlparse==0.3.1
django-cors-middleware==1.5.0
django-cors-headers==3.2.1
PyJWT==1.7.1
django-extensions==2.2.8
phonenumberslite==8.12.10
requests==2.25.1
djangorestframework-simplejwt==4.0.0
